// src/types.ts
  export interface Flag {
    identifier: string;
    text: string;
    color: string;
  }

  export type PlayersQueryParams = {
    page?: string;
    q?: {
      name: string;
    };
  };

  export type Players = Player[];

  export interface Player {
    id: string;
    display_name: string;
    region: string;
    rank: number;
    rank_active: number;
    is_active: boolean;
    rank_title: string;
    rank_image: string;
    is_ranked: boolean;
    mmr: number;
    flags: Flag[];
  }